fix: change the tests to use the test address and fake server
The previous tests used the blockchain directly. This reduced repeatability because all tests that modified the blockchain data would need to be edited after every run.
Also it introduced external failure points due to issues in the grpc-server or the blockchain connection

Now the tests use the python fake server. It verifies the request body is correct based on config.json and returns canned responses